+++ local/modules/aniart.main/lib/observers/bitrixobserver.php
@@ -17,7 +17,7 @@ class BitrixObserver
 
     public static function onEpilog()
     {
-        self::setSeoParams();
+
     }
 
     protected static function initAdditionalUserParams()
@@ -32,21 +32,5 @@ class BitrixObserver
                 $_SESSION['SESS_AUTH']['PERSONAL_MOBILE'] = $userData['PERSONAL_MOBILE'];
             }
         }
-    }
-
-    protected static function setSeoParams()
-    {
-	    /**
-	     * @var SmartSeo $smartSeo
-	     */
-	    $smartSeo = app('SmartSeo');
-	    if($smartSeo->isPageFound()) {
-		    $pageMeta = $smartSeo->getPageMeta();
-		    seo()->setPageTitle($pageMeta['page_title'], true);
-		    seo()->setMetaTitle($pageMeta['meta_title'], true);
-		    seo()->setKeywords($pageMeta['keywords'], true);
-		    seo()->setDescription($pageMeta['description'], true);
-	    }
-        seo()->process();
     }
 }
