--- local/modules/aniart.main/lib/repositories/seopagesrepository.php
@@ -1,13 +1,0 @@
-<?php
-
-
-namespace Aniart\Main\Repositories;
-
-
-class SeoPagesRepository extends AbstractHLBlockElementsRepository
-{
-	public function newInstance(array $fields)
-	{
-		return app('SeoPage', array($fields));
-	}
-}
