--- local/modules/aniart.main/lib/smartseo/expressions/phraseexpression.php
@@ -1,10 +1,0 @@
-<?php 
-
-namespace Aniart\Main\SmartSeo\Expressions;
-
-use Aniart\Main\SmartSeo;
-
-class PhraseExpression extends NonTerminalExpression
-{
-
-}
