--- local/modules/aniart.main/lib/smartseo/expressions/terminalexpression.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Aniart\Main\SmartSeo\Expressions;
-
-use Aniart\Main\SmartSeo;
-
-class TerminalExpression extends AbstractExpression
-{
-
-}
