--- local/modules/aniart.main/lib/smartseo/interfaces/interpretedinterface.php
@@ -1,9 +1,0 @@
-<?php 
-
-namespace Aniart\Main\SmartSeo\Interfaces;
-
-interface InterpretedInterface
-{
-	public function interpret(InterpreterInterface $interpreter);
-}
-?>
