--- local/modules/aniart.main/lib/smartseo/interfaces/pagesrepositoryinterface.php
@@ -1,9 +1,0 @@
-<?php 
-
-namespace Aniart\Main\SmartSeo\Interfaces;
-
-interface PagesRepositoryInterface
-{
-	public function getByCurrentUri();
-	public function getByUri($uri);
-}
