--- local/modules/aniart.main/lib/smartseo/interfaces/parserinterface.php
@@ -1,10 +1,0 @@
-<?php 
-
-namespace Aniart\Main\SmartSeo\Interfaces;
-
-interface ParserInterface
-{
-	public function __construct($data = null);
-	public function setData($data);
-	public function parse();
-}
