+++ local/modules/aniart.main/misc.php
@@ -33,20 +33,3 @@ function app($abstract = null, $params = array()) {
 	}
 	return $app->make($abstract, $params);
 }
-
-function seo($paramName = null, $paramValue = null, $overwrite = false)
-{
-    /**
-     * @var \Aniart\Main\Seo\SeoParamsCollector $seo;
-     */
-    $seo = app('SeoParamsCollector');
-    if(!is_null($paramName)){
-        if(is_null($paramValue)){
-            return $seo->getParamValue($paramName);
-        }
-        else{
-            return $seo->setParamsValue($paramName, $paramValue, $overwrite);
-        }
-    }
-    return $seo;
-}
