+++ local/modules/aniart.main/vars.php
@@ -1,3 +1,3 @@
 <?php
 
-define('HL_SEO_PAGES_ID', 9);
+define('IB_PRODUCTS_ID', 2);
