+++ local/modules/aniart.seo/events.php
@@ -1,0 +1,9 @@
+<?php
+/**
+ * Регистрация обработчиков системный событий
+ */
+use Bitrix\Main\EventManager;
+$eventManager = EventManager::getInstance();
+
+$eventManager->addEventHandler('main', 'OnEpilog', ['\Aniart\Seo\Observers\BitrixObserver', 'onEpilog']);
+$eventManager->addEventHandler('main', 'OnProlog', ['\Aniart\Seo\Observers\BitrixObserver', 'onProlog']);
