+++ local/modules/aniart.seo/lib/smartseo/expressions/fieldexpression.php
@@ -1,0 +1,10 @@
+<?php 
+
+namespace Aniart\Seo\SmartSeo\Expressions;
+
+use Aniart\Seo\SmartSeo;
+
+class FieldExpression extends NonTerminalExpression
+{
+	
+}
