+++ local/modules/aniart.seo/lib/smartseo/expressions/functionexpression.php
@@ -1,0 +1,21 @@
+<?php 
+
+namespace Aniart\Seo\SmartSeo\Expressions;
+
+use Aniart\Seo\SmartSeo;
+
+class FunctionExpression extends TerminalExpression
+{
+	protected $arguments;
+
+	public function setArguments(array $arguments)
+	{
+		$this->arguments = $arguments;
+		return $this;
+	}
+	
+	public function getArguments()
+	{
+		return $this->arguments;
+	}
+}
