+++ local/modules/aniart.seo/lib/smartseo/expressions/terminalexpression.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Aniart\Seo\SmartSeo\Expressions;
+
+use Aniart\Seo\SmartSeo;
+
+class TerminalExpression extends AbstractExpression
+{
+
+}
