+++ local/modules/aniart.seo/lib/smartseo/interfaces/interpretedinterface.php
@@ -1,0 +1,9 @@
+<?php 
+
+namespace Aniart\Seo\SmartSeo\Interfaces;
+
+interface InterpretedInterface
+{
+	public function interpret(InterpreterInterface $interpreter);
+}
+?>
