+++ local/modules/aniart.seo/lib/smartseo/interfaces/pagesrepositoryinterface.php
@@ -1,0 +1,9 @@
+<?php 
+
+namespace Aniart\Seo\SmartSeo\Interfaces;
+
+interface PagesRepositoryInterface
+{
+	public function getByCurrentUri();
+	public function getByUri($uri);
+}
