+++ local/modules/aniart.seo/lib/smartseo/interfaces/parserinterface.php
@@ -1,0 +1,10 @@
+<?php 
+
+namespace Aniart\Seo\SmartSeo\Interfaces;
+
+interface ParserInterface
+{
+	public function __construct($data = null);
+	public function setData($data);
+	public function parse();
+}
