+++ local/modules/aniart.seo/lib/smartseo/seo.json
