+++ local/modules/aniart.seo/utils.php
@@ -1,0 +1,1 @@
+<?php
