+++ local/modules/aniart.seo/vars.php
@@ -1,0 +1,3 @@
+<?
+define('HL_SEO_PAGES_ID', 19);
+
