+++ local/php_interface/init.php
@@ -1,2 +1,3 @@
 <?php
 \Bitrix\Main\Loader::includeModule('aniart.main');
+\Bitrix\Main\Loader::includeModule('aniart.seo');
